+++ src/builder.ts
@@ -1,4 +1,4 @@
-import { Database, Table } from './schema'
+import { Database, Table } from './database'
 import { Condition } from './models'
 import { Binder, BinderStore } from './binder'
 import { ASTERISK, Distinct, All } from './singletoneConstants'
@@ -15,11 +15,12 @@ import {
 import { OrderByItemInfo } from './orderBy'
 import { SelectItemInfo } from './select'
 import { BuilderOption, fillUndefinedOptionsWithDefault } from './option'
+import { MoreThanOneDistinctOrAllError } from './errors'
 
 export type BuilderData = {
-  dbSchema: Database,
+  database: Database,
   option: BuilderOption,
-  /** Below data used to generate SQL statment */
+  /** Below data used to generate SQL statement */
   selectItemInfos: SelectItemInfo[],
   //TODO: make table "FromItemInfo" array
   table?: Table,
@@ -37,7 +38,7 @@ export class Builder {
 
   constructor(database: Database, option?: BuilderOption) {
     this.data = {
-      dbSchema: database,
+      database: database,
       table: undefined,
       selectItemInfos: [],
       distinct: '',
@@ -56,7 +57,7 @@ export class Builder {
       if (items.length <= 1) throw new Error('Select step must have at least one parameter after DISTINCT')
       this.rootStep.cleanUp()
       items.shift() //remove first item the DISTINCT item
-      //TODO: throw error if items contain another DISTINCT or ALL
+      Builder.throwIfMoreThanOneDistinctOrAll(items)
       const newItems = items as unknown[] as (SelectItemInfo|SelectItem|PrimitiveType)[]
       return this.rootStep.selectDistinct(...newItems)
     }
@@ -64,14 +65,14 @@ export class Builder {
     if (items[0] instanceof All) {
       this.rootStep.cleanUp()
       items.shift() //remove first item the ALL item
-      //TODO: throw error if items contain another ALL or DISTINCT
-      const newItems = items as unknown[] as (SelectItemInfo|SelectItem|PrimitiveType)[]
+      Builder.throwIfMoreThanOneDistinctOrAll(items)
+      const newItems = items as (SelectItemInfo|SelectItem|PrimitiveType)[]
       return this.rootStep.selectAll(...newItems)
     }
 
     this.rootStep.cleanUp()
-    //TODO: throw error if items contain any ALL or DISTINCT
-    const newItems = items as unknown[] as (SelectItemInfo|SelectItem|PrimitiveType)[]
+    Builder.throwIfMoreThanOneDistinctOrAll(items)
+    const newItems = items as (SelectItemInfo|SelectItem|PrimitiveType)[]
     return this.rootStep.select(...newItems)
   }
 
@@ -91,5 +92,12 @@ export class Builder {
     //Note: the cleanup needed as there is only one "select" step in the chain that we start with
     this.rootStep.cleanUp()
     return this.rootStep.select(ASTERISK).from(table)
+  }
+
+  private static throwIfMoreThanOneDistinctOrAll(items: (Distinct|All|SelectItemInfo|SelectItem|PrimitiveType)[]) {
+    items.forEach(it => {
+      if (it instanceof Distinct || it instanceof All)
+        throw new MoreThanOneDistinctOrAllError('You can not have more than one DISTINCT or ALL')
+    })
   }
 }