+++ src/index.ts
@@ -12,14 +12,16 @@ export {
   ColumnNotFoundError,
   TableNotFoundError,
   InvalidExpressionError,
+  MoreThanOneDistinctOrAllError,
 } from './errors'
 export {
   Database,
+  Schema,
   Table,
   BooleanColumn,
   NumberColumn,
   TextColumn,
-} from './schema'
+} from './database'
 export {
   OrderByDirection,
   DIRECTION_NOT_EXIST,