+++ src/steps.ts
@@ -1,5 +1,5 @@
 import { Condition, Expression, PostgresBinder } from './models'
-import { Column, Table } from './schema'
+import { Column, Table } from './database'
 import { ColumnNotFoundError, TableNotFoundError } from './errors'
 import { BuilderData } from './builder'
 import { All, Asterisk } from './singletoneConstants'
@@ -252,7 +252,7 @@ export class Step implements BaseStep, RootStep, SelectStep, FromStep, AndStep,
   }
 
   private throwIfTableNotInDb(table: Table) {
-    if (!this.data.dbSchema.isTableExist(table))
+    if (!this.data.database.isTableExist(table))
       throw new TableNotFoundError(`Table: ${table} not found`)
   }
 
@@ -268,22 +268,9 @@ export class Step implements BaseStep, RootStep, SelectStep, FromStep, AndStep,
         continue
       }
       // item is Column from here
-      // TODO: move search function into database model
-      let found = false
-      //@formatter:off
-      COL:
-      //TODO: filter only the table in the current query
-      for (const table of this.data.dbSchema.getTables()) {
-        for (const col of table.getColumns()) {
-          if (item === col) {
-            found = true
-            break COL
-          }
-        }
+      if (!this.data.database.isColumnExist(item)) {
+        throw new ColumnNotFoundError(`Column: ${item} not found in database`)
       }
-      //@formatter:on
-      if (!found)
-        throw new ColumnNotFoundError(`Column: ${item} not found`)
     }
   }