+++ test/diffOptions.test.ts
@@ -1,19 +1,16 @@
-import { Builder, Database, Table, TextColumn } from '../src'
+import { Builder } from '../src'
+import {
+  database,
+  table,
+  column1,
+  column2,
+} from './database'
 
 describe('test Options', () => {
-  const column1 = new TextColumn('col1')
-  const column2 = new TextColumn('col2')
-
-  const table = new Table(
-    'testTable',
-    [column1, column2],
-  )
-  const db = new Database([table], 1)
-
   describe('test Semicolon Option', () => {
-    const sqlWithoutSemicolon = new Builder(db, { useSemicolonAtTheEnd: false })
-    const sqlWithSemicolon = new Builder(db, { useSemicolonAtTheEnd: true })
-    const sqlDefault = new Builder(db)
+    const sqlWithoutSemicolon = new Builder(database, { useSemicolonAtTheEnd: false })
+    const sqlWithSemicolon = new Builder(database, { useSemicolonAtTheEnd: true })
+    const sqlDefault = new Builder(database)
     it('Produces [SELECT 1 FROM "testTable"] without semicolon', () => {
       const actual = sqlWithoutSemicolon
         .select(1)
@@ -43,10 +40,10 @@ describe('test Options', () => {
   })
 
   describe('test OrderBy ASC Option', () => {
-    const sqlAlways = new Builder(db, { addAscAfterOrderByItem: 'always' })
-    const sqlNever = new Builder(db, { addAscAfterOrderByItem: 'never' })
-    const sqlWhenMentioned = new Builder(db, { addAscAfterOrderByItem: 'when mentioned' })
-    const sqlDefault = new Builder(db)
+    const sqlAlways = new Builder(database, { addAscAfterOrderByItem: 'always' })
+    const sqlNever = new Builder(database, { addAscAfterOrderByItem: 'never' })
+    const sqlWhenMentioned = new Builder(database, { addAscAfterOrderByItem: 'when mentioned' })
+    const sqlDefault = new Builder(database)
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" ASC;] option(always)', () => {
       const actual = sqlAlways
         .select(column1)
@@ -129,10 +126,10 @@ describe('test Options', () => {
   })
 
   describe('test OrderBy NULLS LAST Option', () => {
-    const sqlAlways = new Builder(db, { addNullsLastAfterOrderByItem: 'always' })
-    const sqlNever = new Builder(db, { addNullsLastAfterOrderByItem: 'never' })
-    const sqlWhenMentioned = new Builder(db, { addNullsLastAfterOrderByItem: 'when mentioned' })
-    const sqlDefault = new Builder(db)
+    const sqlAlways = new Builder(database, { addNullsLastAfterOrderByItem: 'always' })
+    const sqlNever = new Builder(database, { addNullsLastAfterOrderByItem: 'never' })
+    const sqlWhenMentioned = new Builder(database, { addNullsLastAfterOrderByItem: 'when mentioned' })
+    const sqlDefault = new Builder(database)
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" NULLS LAST;] option(always)', () => {
       const actual = sqlAlways
         .select(column1)
@@ -215,9 +212,9 @@ describe('test Options', () => {
   })
 
   describe('test SelectItems AS Option', () => {
-    const sqlAlways = new Builder(db, { addAsBeforeColumnAlias: 'always' })
-    const sqlNever = new Builder(db, { addAsBeforeColumnAlias: 'never' })
-    const sqlDefault = new Builder(db)
+    const sqlAlways = new Builder(database, { addAsBeforeColumnAlias: 'always' })
+    const sqlNever = new Builder(database, { addAsBeforeColumnAlias: 'never' })
+    const sqlDefault = new Builder(database)
     it('Produces [SELECT "col1" AS "C1" FROM "testTable";] option(always)', () => {
       const actual = sqlAlways
         .select(column1.as('C1'))