+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
-## [0.8.0]  2022-02-16
+## [0.8.1]  2022-02-25
+### Added
+- in test folder add schema in separate folder
+### Breaking Changes
+- restructure Database schema classes
+  - move version no to Database 
+  - rename class Database to Schema
+  - tables now belong to Schema
+  - create class Database that contain several schemas
+  - construct database schema by object
+
+## [0.8.0]  2022-02-17
 ### Added
 - Add LIMIT and OFFSET steps
 - Add LIMIT$ and OFFSET& for binds values